Remove schema argument from RowType methods
The extra schema argument was added in the beginning for the "alter schema".
but "alter schema" works at a different level and each Row will have its own schema.
if you want to query a Row with a different schema you must use ProjectRow() or similar,
which is the read-path followed by the client/server.

Change-Id: Ia0d1cb921af776e87ddad978e32c37000ba003ca
Reviewed-on: http://gerrit.ent.cloudera.com:8080/460
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>