Change TabletSuperBlockPB.last_durable_mrs_id to int64 from uint64
We need to distinguish when there hasn't been any mrs flush from the
first one. TabletMetadata already set the last_durable_mrs_id to -1
but the var in TabletSuperBlockPB was uint64 and not int64 which would
cause trouble when adding the current meta to the log segment header
(i.e when extacting the current TabletSuperBlockPB without UpdateAndFlush).

Also added a proper constant (kNoDurableMrs) to which last_durable_mrs_id
is set when a new TabletMetadata is created.

Change-Id: I9bc826e5b1bea85c588e97d5f5addd207fa922af
Reviewed-on: http://gerrit.ent.cloudera.com:8080/462
Tested-by: jenkins
Reviewed-by: Matteo Bertozzi <matteo.bertozzi@cloudera.com>