Adds WAL/SM Log (writer) and LogReader
The Log class is the Log writer. It pre-allocates space in disk
for a segment and writes to that segment until a pre-set max
size is reached. When this happens the Log rolls to a new
segment, The header is kept up to date with the id of the
last message that was added, the latest metadata, etc.. This
eases in creating new segments and later, if needed, will allow
to query for the most recent entry id/meta/config. The Log
writer is capable of continuing an existing Log (e.g. if a
node crashed and after performing recovery, we might use a
portion of the old log to catch up other nodes) and is able to
perfom GC on the list of previous segments. Segment GC is not
automated and has to be called explicitely.

The LogReader, on open, reads a set of segments and parses the
headers (but not the log entries). LogReader::ReadEntries can be
called to read the set of entries in a segment. Later this might be
changed to receive an entries processor of some kind, instead of
simply returning the whole list. Although the reader will be later
used by the LogPlayer it does not replay entries itself.

Performance wise, internally, the Log writer still needs some
work, but API wise it should be very close (if not exactly) what
is required for the other components that depend on Log.

Change-Id: I49462d49bc1271fafa65f6694ca94cb94e26c7d3
Reviewed-on: http://gerrit.ent.cloudera.com:8080/501
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>