Initial work on TS<->Master protocol and TS Manager
This lays some initial groundwork in the master for receiving
tablet heartbeats, and the initial RPC protocol for this interaction.

This commit does not yet hook up anything on the Tablet Server
side, but already has grown somewhat large, so I think it's worth
committing with just the simple unit tests.

Change-Id: Ic412516ac3f3d587498ef23b386734b69705b079
Reviewed-on: http://gerrit.ent.cloudera.com:8080/508
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>