Add a Slice::CheckSize static method and move most impl to slice.cc.
This patch adds a Slice::CheckSize() static method to slice that checks
that the slice is not null and that the slice's size is the expected one.
The method returns Status::Corruption if any check failed and includes
the slices ToDebugString().

This also moves some of the implementation of slice.h to slice.cc as
otherwise a circular dependency would occur between status.h and slice.h.

Change-Id: I77b9dc05a97ecbdcf9a135558a5331f646dd0e87
Reviewed-on: http://gerrit.ent.cloudera.com:8080/511
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>