Multi-Task/Consensus based write path for TabletServer.
This patch introduces the remaining functionality for Multi-Task
(as in Prepare and Apply), asynchronous execution to TabletServer.
Writes are now split in two phases divided by a barrier. The first
phase executes Prepare and Log, the second phase executes Apply
and Commit. A detailed diagram and explanation on how writes are
now executed can be found in tablet_peer.h.

Change-Id: I7f7ada444095e0b4c17ccd63b83d6bae907e7201
Reviewed-on: http://gerrit.ent.cloudera.com:8080/531
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>