Add PreparedRowWrite and Insert/MutateUnlocked.
This patch is another milestone in the move to multi task,
consensus based, inserts/mutates.

TransactionContext in enhanced to actually hold all the
locks, the mvcc transaction and everything else related to
a transaction. All of these are destroyed on the correct order
on dctor and Reset().

A new construct that holds the row or row_key/mutation and the
row lock was added, the PreparedRowWrite. New tablet methods
were created that receive PreparedRowWrite instead of the raw
row or row_key/mutation. The original methods delegate to the
new ones. For now delegation is useful for several reasons like
making sure that everything works as before without changing
tests or illustrating how things will work on PrepareTask.

Change-Id: If8e8aea92e9a216d80ad64f7cc97fb3c5786f96c
Reviewed-on: http://gerrit.ent.cloudera.com:8080/516
Tested-by: David Alves <david.alves@cloudera.com>
Reviewed-by: Todd Lipcon <todd@cloudera.com>