KUDU-11. Fix reactor thread startup to not be racy
The code which started the reactor thread started the thread
and then assigned the thread pointer into a member variable.
But, many of the functions run by the thread contain assertions
that the member variable points to the current thread -- these
assertions could fail if the thread started quickly and then
the new thread ran for a while before the *starting* thread
did the assignment of the member variable.

This commit adds a utility function which can be used to start
a thread in a blocked (spinning) state, assign the thread to
a variable, and then allow the thread to continue.

Change-Id: I5865653ce076753cfdd783cce60e44cabf397f21
Reviewed-on: http://gerrit.ent.cloudera.com:8080/524
Tested-by: jenkins
Reviewed-by: David Alves <david.alves@cloudera.com>