KUDU-10. Fix UBSAN failures and other crashes on ReactorTask dtor
Fixes a bug seen where, if there were pending tasks on the ReactorTask
when it shuts down, it would either trigger a boost assertion failure
or a UBSAN error.

The issue was that each task's Abort() method was called before removing
it from the list. The Abort() method typically will delete the task
itself, so then trying to follow that list element's 'next' pointer
in the intrusive list would reference now-deleted memory.

Change-Id: Ia8c2559953441ea03be9031fa713f83a1c692c81
Reviewed-on: http://gerrit.ent.cloudera.com:8080/525
Tested-by: jenkins
Reviewed-by: David Alves <david.alves@cloudera.com>