Make fsync()'ing on all calls to Log::Append() optional.
This patch makes calling fsync() on all calls to Append() optional
through a flag, and also makes the default 'false', i,e, not to
explitcitly fsync. This patch also changes the behavior when fsync
in on to fsync() all messages and not only COMMIT messages. With this
change, when fsync() is on, the Log will comply with the durability
requirements for distributed consensus.

Change-Id: I649f278602bffffe3c38a13036857703c7dc7df0
Reviewed-on: http://gerrit.ent.cloudera.com:8080/534
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>