KUDU-9. Switch uses of Read() to ReadFully()
The read() syscall may spuriously return short reads. We had
many places relying on Read() returning the full number of requested
bytes. This adds a ReadFully wrapper which loops the read call until
the correct number of bytes are returned, or returns an IOError
if it hits EOF before the full number of bytes.

Change-Id: I2967791211333e421e7d9a5ca7957bfaa3383c84
Reviewed-on: http://gerrit.ent.cloudera.com:8080/555
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>