Introduce a ServerBase class which master and tablet server inherit from.
This serves to also introduce the web server into the master, while
combining some common code.

This will also be a nice place to put other server-common infrastructure
like a metrics registry, and may make it easier to share some test utility
code (eg starting a server on port 0 and then fetching back which ports
it actually bound to in the context of a minicluster).

Change-Id: I788d51f68787dd17ccc19853893768e260f47df0
Reviewed-on: http://gerrit.ent.cloudera.com:8080/558
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>