Rename "MasterServer" to "Master"
This terminology is a little simpler and matches how we refer to it
in casual conversation.

Also renames the executables for the master and the tablet server to
"kudu-master" and "kudu-tablet_server", to avoid naming conflicts
between the 'master' CMake target (the static library with the master
code) and the 'kudu-master' CMake target (the executable daemon).

Change-Id: I43d3f623212e6dca8f1f9e6fa71da53d7f2c1f65
Reviewed-on: http://gerrit.ent.cloudera.com:8080/581
Tested-by: jenkins
Reviewed-by: Matteo Bertozzi <matteo.bertozzi@cloudera.com>