KUDU-20. Fix RPC behavior when queue is full or shut down
This fixes an issue where, if the service queue is shut down while calls
are still being received, then the failure to enqueue the call would
call RespondFailure, which would then fire a DCHECK assertion because
a response would be generated from the reactor thread instead of the
service thread.

The fix for the crash is to remove a few DCHECKs. In fixing it, we also
noticed that the failure response was incorrectly attributing this
failed RPC to backpressure, when in fact it could also be due to the
service shutting down.

Added mt-rpc-test which starts a few threads hammering a server, and
then shuts down the server. Another similar test does that with a bunch
of concurrent calls and makes the queue back up by not removing incoming
calls from the service queue, provoking the "queue out of space" problem.

Also added a few new Status message types to make it easier to
differentiate various types of error conditions and cleaned up a couple
of tests that depended on RPC errors being RuntimeErrors instead of
RemoteErrors.

Change-Id: I00aedc5cb2a83b18c6c33619e3eeaa573948cd9c
Reviewed-on: http://gerrit.ent.cloudera.com:8080/588
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>