rle: support bounding how many values each GetNextRun call returns
This changes the API for GetNextRun so that the user may specify a maximum
number of values to consume at once. This is in order to help address a
performance problem with NULL handling (see also KUDU-23). The previous API
here caused an issue when processing NULL columns since each row block to be
read was likely to be shorter than a NULL run, so the NULL decoder would end up
advancing past the intended row on every block, and thus have to be re-seeked
constantly.

This commit does not itself address the issue in the CFile read path -- just
introduces the API to be used by a later commit.

Change-Id: I5b883cf2d7bfed2a21d1759b43f19cdd7e9613d7
Reviewed-on: http://gerrit.ent.cloudera.com:8080/614
Tested-by: jenkins
Reviewed-by: Matteo Bertozzi <matteo.bertozzi@cloudera.com>