metadata: store Tablet schema within TabletMetadata
This is needed to re-open the tablets given only the super block.
It still needs some further work to integrate with AlterSchema, which
is concurrently being worked on, but this should hopefully unblock
my other patch which persists the set of tablets on disk and re-opens
them after TS restart.

This also does some cleanup on the construction and loading of TabletMetadata
to be done through factory methods, in order to avoid exposing half-constructed
TabletMetadata objects. The Tablet construction is also now amended to only
need a TabletMetadata and no longer take its own copy of the schema.

Change-Id: Idb2e3b96cad0e0904060d7a1c201536d6fb935b0
Reviewed-on: http://gerrit.ent.cloudera.com:8080/633
Tested-by: jenkins
Reviewed-by: Matteo Bertozzi <matteo.bertozzi@cloudera.com>