Scripts for continuous benchmarking and test fixups
First this change fixes a couple of tests to make them easier to parse
or just to make them do more.
It also adds python scripts to get data in and out of MySQL.
A new R script is added to map all the benchmarks the same way.
mt-tablet-test-graph.R is modified to be run in batch.

Change-Id: If7cc8fb81008160db1747d24e4fd1e9a9d92fb85
Reviewed-on: http://gerrit.ent.cloudera.com:8080/830
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>