ts_tablet_manager: generate full and incremental tablet reports
This includes (but does not integrate) the code in the TSTabletManager to
track which tablets have been changed since the last incremental report,
and generate a TabletReportPB to send to the master.

Currently, the only changes which can be reported are additions and removals,
but the same infrastructure should work to track any state changes as well.

Change-Id: I381c7338b51fd38bff42a7568a3de84768db7890
Reviewed-on: http://gerrit.ent.cloudera.com:8080/649
Reviewed-by: Matteo Bertozzi <matteo.bertozzi@cloudera.com>
Tested-by: jenkins