Move generation of instance ID into ServerBase
This patch also changes the sequence number to be time-based instead of
constant. This helps with some integration tests where we rely on
restarting a tablet server and assume that the new instance has a higher
sequence number (and thus can re-register).

The timer used isn't an actual wall-clock time, though, so this is
useless in real life. We will need to come back to this relatively
soon, but this unblocks the above-mentioned tests.

Change-Id: I930fcff0c8b2a76dbe6b43d2d1ee7c3ff1d82f1e
Reviewed-on: http://gerrit.ent.cloudera.com:8080/653
Tested-by: jenkins
Reviewed-by: Michael Stack <stack@cloudera.com>
Reviewed-by: David Alves <david.alves@cloudera.com>