Fix component lock release bug
Currently we're releasing both the row locks and the component lock on
"early lock release". This is a serious bug as it allows flushes/compactions
to creep in. This patch fixes this by only releasing the row locks on ELR
and retaining the component lock until the mvcc tx is committed.

NOTE: Testing for this will follow in the updated mt-tablet-server test patch.

Change-Id: I12c49d374d31bfe48a9f506ad3fb9a39791867a3
Reviewed-on: http://gerrit.ent.cloudera.com:8080/658
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>