Add Column IDs to support rename and delete/re-add
To support column rename or delete/re-add, the RowChangeList
must be serialized with IDs instead of column indexes.

The IDs assignment will be done during AlterTable() on the server side,
with the Master projecting the IDs based on the client operations:
addColumn(), deleteColumn(), renameColumn().

The client doesn't know anything about the IDs, and it only sends
the schemas with column name and type.

Change-Id: I1046589d56d32f3e4a6e1532eca626cb71d90edf
Reviewed-on: http://gerrit.ent.cloudera.com:8080/673
Tested-by: jenkins
Reviewed-by: David Alves <david.alves@cloudera.com>