Make TransactionContext take the row key for missed delta mutations
Currently missed deltas are added to a transaction result as a row_idx
and a changelist. However using just this information on replay is
tricky, as it would require to inspect all delta trackers for the place
to put it (akin to what compaction.cc does in ReupdateMissedDeltas).
This maily because the common mutate path cannot be used since we don't
have the mutation's row key.

This patch changes this to add the row key to the missed delta transaction
result, allowing to use the usual method for mutations, i.e. Tablet::Mutate().

Change-Id: If1930460675e812d52c6457bf4f013cca19ce6a8
Reviewed-on: http://gerrit.ent.cloudera.com:8080/741
Tested-by: jenkins
Reviewed-by: Matteo Bertozzi <matteo.bertozzi@cloudera.com>