Allow LogReader to open segments from the recovery dir
On recovery segments are moved to a recovery directory so that
a new log can be created on replay. This adds an Open() overload
to LogReader that allows to read from the recovery dir.

Also added a log entry to Log to make clear whether fsync() is on
or off.

Change-Id: Id3919d559e66fdf258f9ce02d22facf875296128
Reviewed-on: http://gerrit.ent.cloudera.com:8080/744
Tested-by: jenkins
Reviewed-by: Matteo Bertozzi <matteo.bertozzi@cloudera.com>