Implement major delta compaction
Implements major delta compaction:

1) Creates a method that iterates over all deltas in a given
prepared batch of rows in DeltaIterator and appends these deltas
to specified vector.

2) Uses the above method and RowSetColumnUpdater from
a previous revision apply deltas to specified columns of a
DiskRowSet, writing the new DiskRowSet and the deltas for
other columns to disk.

Remaining items (for future revisions):

1) Keep track of which columns have the most updates so that we know
which columns can be updated

2) Possibly handle DeltaMemStore and MemRowSet for delta compaction?

Change-Id: I5e73720699d38e56025703bb66604d8d8040552d
Reviewed-on: http://gerrit.ent.cloudera.com:8080/805
Tested-by: jenkins
Reviewed-by: Matteo Bertozzi <matteo.bertozzi@cloudera.com>