Add the SchemaBuilder helper
The schema builder will be used on master side AlterTable()
to create a new schema based on the operations requested by
the user (addColumn, removeColumn, renameColumn).

The schema builder is responsible to handle the ID mappings.
IDs are monotonically increasing and can't be reused. To be able to reuse
the IDs you have to make sure that there are no delta-files using them.
that means a full compaction on each Tablet Server is required.

Change-Id: I98876ca2449317c6285951ff3b87210de1579d44
Reviewed-on: http://gerrit.ent.cloudera.com:8080/754
Reviewed-by: David Alves <david.alves@cloudera.com>
Tested-by: jenkins