Allow to retrieve the OpId of the last Log entry
In order to continue a log, catch up with other nodes through RPC
or even just to perform sanity checks we need to know the OpId of
the last log entry. The new method is currently getting the OpId of
the next segment's header but that might change in the future.

Change-Id: Ibb182893a5f805fb58fd330752f81fa3e764ddf7
Reviewed-on: http://gerrit.ent.cloudera.com:8080/790
Tested-by: jenkins
Reviewed-by: Matteo Bertozzi <matteo.bertozzi@cloudera.com>