master: Add GetTabletLocations RPC to fetch current locations of tablets
This will eventually be fleshed out to be a lot more complete, handling
reporting which replica is the leader, etc, but this is sufficient for now
to start plugging into the client.

This patch also removes the TabletServerPB from metadata.proto, since
the tablet locations are no longer part of the persistent metadata in the
current iteration of the design. Changes QuorumPB to reference the TS
UUID rather than the TS host/port, since quorums should be able to
survive a restart even if the hostnames or ports are reconfigured.

Change-Id: Ibe1b10ff15ce4597d26303299826aaa9353deea3
Reviewed-on: http://gerrit.ent.cloudera.com:8080/823
Tested-by: jenkins
Reviewed-by: David Alves <david.alves@cloudera.com>