Split tserver.proto into a service definition and regular PBs
The goal is to break a cyclic dependency where consensus depends on the
PBs defined in tserver.proto and the tablet server service definition in turn
depends on consensus.proto (since it will be the RPC interface for consensus).
The solution is to create a new tserver_service.proto with only the service
which depends on consensus.proto and tserver.proto.

Change-Id: Ief154b94102d591e376b6d679aa9ea72538b067e
Reviewed-on: http://gerrit.ent.cloudera.com:8080/815
Tested-by: jenkins
Reviewed-by: Matteo Bertozzi <matteo.bertozzi@cloudera.com>