Add Bootstrap/Recovery and related tests
This patch adds a complete implementation of recovery and a bunch
of tests to make sure it works.

TSTabletManager now uses TabletBootstrap to replay log entries and
rebuild soft state on opening a tablet that already has blocks on
disk. There are some initial pointers as to how this would work with
dist/quorum based recovery.

Still to come is the multi-process test, but the recovery implementation
is mostly finished.

Change-Id: I86ad2972b81a9eb746867a09db0f643c51d7f82c
Reviewed-on: http://gerrit.ent.cloudera.com:8080/751
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>