Make client look up tablet locations by RPCing to master
This patch sets up some skeleton code for the rich client.
The client now maintains a map containing all of the tablet
servers, as well as a map containing the tablets. When it
needs to look up a tablet location, it first consults this
map, and if necessary looks up the tablet by RPCing to the master.

The internal implementation is designed to be async, since many
operations will involve fan-out writes to many servers, etc,
and we don't want to block one of the operations in the fan-out
because of an unrelated DNS lookup or master RPC. The async-ness
isn't yet fully plumbed through, but the intention is that the
synchronous client API will be a shallow wrapper around an async
one.

There are a number of TODOs in this code which will be addressed
in subsequent commits.

This patch also changes the existing client-test to set up a minicluster
instead of directly RPCing to the master.

Change-Id: Ia3b784890c259f1cee81abe5d6ee96f2c7d390ee
Reviewed-on: http://gerrit.ent.cloudera.com:8080/838
Tested-by: jenkins
Reviewed-by: Michael Percy <mpercy@cloudera.com>