KUDU-35. TS registration should not include wildcards
The various GetBoundAddresses() calls will return the wildcard (0.0.0.0)
IP address, which is obviously not useful for other nodes to remotely
contact the local node. So, when registering, the TS should replace
wildcard addresses with the hostname of the machine.

Another option would have been to enumerate all local interfaces and send
all IPs separately, but chose to do the hostname-based solution as it
matches the default behavior of other Hadoop-stack software (HDFS, Impala, etc)

Change-Id: Ic43e11a9b808e41b456513f48464327135fe2041
Reviewed-on: http://gerrit.ent.cloudera.com:8080/841
Tested-by: jenkins
Reviewed-by: Matteo Bertozzi <matteo.bertozzi@cloudera.com>