Fix a bug where the LogReader would fail if a log segment wasn't ftruncate()d
Because log segments are pre-allocated, if the server crashes without making the
ftruncate() call, on replay the LogReader would read an entry length of 0 and assume
that the log was corrupted. This patch changes the assumption, making the reader
simply assume it is done reading that log segment when it finds an entry of length 0.
A related test was added to log-test.cc.

Change-Id: I8fe354fdbcdd849d032c55ae3f4afe7c63381aba
Reviewed-on: http://gerrit.ent.cloudera.com:8080/843
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>