Fix leak, clean up code in Tablet bootstrap
Previously, when matching up Commit messages against their initial
Replicate messages, the original messages were never getting removed
from the entry map. This caused the map to grow to the complete size of
the logs that were read, causing OOMs if reading a lot of logs.

This cleans up the code in TabletBootstrap into smaller functions with
better documented object ownership, and also adds a dump at the end
of the replay process of any REPLICATE messages which were orphaned in
the map, which would have made this problem much more obvious.

Manually tested by using the HEAPPROFILE environment variable and looking
at the memory usage during bootstrap. Without this, it was growing without
bound. With this, it stays constant around 300MB during startup.

Change-Id: I0108a8a4d81ce1a9c7032d0ab960ef3bdcb42746
Reviewed-on: http://gerrit.ent.cloudera.com:8080/856
Tested-by: jenkins
Reviewed-by: David Alves <david.alves@cloudera.com>