KUDU-41. lock_manager: track transaction lock holder, WARN on long acquisition
This adds a field to the lock entry which contains a pointer to the
TransactionContext which holds the lock. A new assertion fires if a
single transaction tries to lock the same row twice (this currently
causes a deadlock, but a crash is preferable).

Additionally, WARNINGs are triggered if lock acquisition takes too
long. This will eventually be hooked into metrics/tracing.

In writing the test, fixed a bug where LmTestThread used the thread
ID of the _constructing_ thread instead of the test thread for its
tid_ parameter.

Change-Id: I5cfd9cab5b2c7d386a8138bc54d191300691386f
Reviewed-on: http://gerrit.ent.cloudera.com:8080/859
Tested-by: jenkins
Reviewed-by: Matteo Bertozzi <matteo.bertozzi@cloudera.com>