schema: change stringification, add ColumnSchema::TypeToString
This changes the stringification of a schema from the pattern
"column_name[type='string' NOT NULL]" to "column_name[string NOT NULL]".
This is a bit more readable.

Also extracted function ColumnSchema::TypeToString which just returns the
type string. This is useful when presenting schemas in other contexts
(eg tabular form on the web interface, in a patch to follow)

Change-Id: Id3b7eeaeba12343c4b2e80d3b5ea059dc363407a
Reviewed-on: http://gerrit.ent.cloudera.com:8080/860
Reviewed-by: Matteo Bertozzi <matteo.bertozzi@cloudera.com>
Tested-by: jenkins