Make Prepare fail when to_mutate row keys != number of mutations
Right now we are assuming that the number of mutations and row keys in the
request are the same, which might not be the case. Best case this allows clients
to add unlimited mutations to a request with only a few (or 1) keys making
requests bigger than they need to be. Worst case this can make us segfault if
num keys > num mutations.

This patch adds a check to make sure the counts are the same and sends back
an error otherwise.

Change-Id: Iacbd1a0de26125f8bd77c944213b1f5479d6cd51
Reviewed-on: http://gerrit.ent.cloudera.com:8080/870
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>