KUDU-54. PredicateEvaluatingIterator needs to respect the selection vector
The selection vector is used by MVCC to filter out values which
should not be seen by the user. When it has done so, we can't count
on the value in the row to be valid -- in fact, in debug builds,
it gets overwritten by "MVCCMVCCMVCC...". The PredicateEvaluatingIterator,
then, was segfaulting trying to access this invalid data as if it were
a valid Slice pointer when evaluating a string range predicate.

Change-Id: I3d0984300ed4003f6c98fc1a7cde8d545372df7c
Reviewed-on: http://gerrit.ent.cloudera.com:8080/905
Tested-by: jenkins
Reviewed-by: Matteo Bertozzi <matteo.bertozzi@cloudera.com>