tablet_service: limit scans to only loop for 500ms
This fixes an issue I saw with SELECT COUNT(*) queries in Impala on larger
datasets. Because there is no actual data being returned, the batch size
threshold never takes effect. This was causing the Scan() RPCs to go for many
seconds, which failed due to the shorter client timeout.

We'd probably want to use the client's RPC timeout to determine this budget but
this is a nice workaround for now.

Tested manually by running the same Impala queries and verifying that they
didn't time out anymore.

Change-Id: Ic5497d388a57177846f34cb1aed91b4b34d996fa
Reviewed-on: http://gerrit.ent.cloudera.com:8080/910
Tested-by: jenkins
Reviewed-by: Jean-Daniel Cryans <jdcryans@cloudera.com>