KUDU-55. tablet_bootstrap: truncate the debug string for error entries
Upon encountering an error, the log replay tried to include the whole protobuf
which caused the problem in the error Status. The log entries can be 100+KB,
so this made the Status object too large to actually log (glog limits log
lines to something like 64KB).

This truncates the protobuf in the Status - we assume that the operator
can use the log-dump utility to find the full message if that's important.

Change-Id: I664596c283624b5ff509c088bcaf54f47d6f0b86
Reviewed-on: http://gerrit.ent.cloudera.com:8080/928
Reviewed-by: David Alves <david.alves@cloudera.com>
Tested-by: jenkins