Milestone 3's Demo
Adds a new executable called ms3_demo that can drive both insert and update load
against a running cluster that has the tpch1 tablet. Inserts are sourced via the
dbgen'd tsv file and the updates are hitting the most recently inserted
rows (configurable). Here's a good command line to use it on a1228,
it will use a 96GB file to insert data:
./build/release/ms3_demo --logtostderr --tpch_demo_updater_threads=5
--tpch_demo_inserter_threads=1 --tpch_max_batch_size=100

Change-Id: I6e82ef89406347352fef999dfa30656cbf4ed91e
Reviewed-on: http://gerrit.ent.cloudera.com:8080/936
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>