Adding an async Java client
This client is leveraging a lot of the AysncHBase code which was modified
to talk to Kudu. Original project: https://github.com/OpenTSDB/asynchbase/
AsyncHBase is "A fully asynchronous, non-blocking, thread-safe,
high-performance HBase client". Those qualities should still be true with
Kudu.

The current dependencies are: Netty, protobuf, log4j, slf4j,
guava, junit, and SU async.

The Maven project's main module is called kudu-client. The
protobuf classes can be generated simply by calling:
./java/kudu-client/dev-support/build-proto.sh

The Java client is currently in kudu.rpc. The RPC and client-side classes
are intertwined so it's not really possible to leave a few in the rpc
package and then some others in a client package. We could also just move
them to the client package.

The client is automatically built and tests are run as part of the
build-support/jenkins/build-and-test.sh script.

There's a unit test that tests all the different flush modes in
TestKuduSession. It runs its own cluster via Runtime.exec and the configurations
can be found in java/kudu-client/src/test/resources/flags. See the new
java/README.md file.

RpcBenchmark can be used to drive load with the client, it's in a separate
module called kudu-client-benchmark. Eventually it will have real benchmarks.

Change-Id: Iff963383cb6e59cb5ad39fab68fafb4ea1b847d9
Reviewed-on: http://gerrit.ent.cloudera.com:8080/1077
Tested-by: jenkins
Reviewed-by: Michael Percy <mpercy@cloudera.com>