Changes necessary to build against stock el6 boost
On the el6 build slaves, we were previously depending on a Cloudera-specific
Boost 1.46 install and symlinks in the system install paths. We should be
able to build on a stock el6 system, which has an older version of boost.
This commit has the necessary changes to get Kudu building on stock el6
without Cloudera's boost backport package installed.

* Checks in boost/uuid library into thirdparty (not included in el6's
  boost)
* Changes includes of boost/thread.hpp to boost/thread/thread.hpp
  because new versions of clang fail to build old versions of the
  boost thread library if the whole library is included.

Change-Id: I57312c1dd5a74a21d3a3803d61ce68c43ccc8db1
Reviewed-on: http://gerrit.ent.cloudera.com:8080/1058
Tested-by: jenkins
Reviewed-by: Jean-Daniel Cryans <jdcryans@cloudera.com>