LockManager: Allow the same txn to acquire the same lock multiple times
This changes the LockManager to return a diffrent lock status (LOCK_ALREADY_AQUIRED)
if the lock was already acquired by the same txn. The caller still keeps a reference
to the lock entry, it just doen't acquire the mutex.

This doesn't completely solve KUDU-47, just the part about no self deadlocks.

Change-Id: Idf80d4259e40cf3cc658cceed9731764dd568fcf
Reviewed-on: http://gerrit.ent.cloudera.com:8080/1099
Tested-by: jenkins
Reviewed-by: Matteo Bertozzi <matteo.bertozzi@cloudera.com>