Add support for ref-counted objects
This is code borrowed from Chromium which performs a similar function as
boost::intrusive_ptr, but a bit nicer in that it also provides a base class
which handles ref counting for you.

This can be slightly faster and smaller than shared_ptr, since the refcount of
the object is held with the object itself. This means that you may avoid a
cache miss, and that allocating a new ref-counted object only requires one call
into the allocator instead of two.

It's also handy in that you can manually manage reference counting in some
places where you have a pointer to the ref-counted object only in function
scope, and don't want to keep around a scoped object in an instance.

CONTRIBUTING.txt is updated to recommend using scoped_refptr over shared_ptr
in new code where possible, explaining the pros/cons.

This commit also imports thread_collision_warner which is a handy tool for
verifying that thread-unsafe structures aren't used from multiple threads at
the same time.

Local changes to implementation:
* Prohibits implicit version from scoped_refptr to a bare pointer while still
  allowing implicit conversion to bool NULL checks.
  (see http://code.google.com/p/chromium/issues/detail?id=110610)
* Slight changes to atomicops to be more compatible with the version from
  Chromium's source tree

Change-Id: Ia66db6ff58ade5ec0bc5e6ab604616d4211563f7
Reviewed-on: http://gerrit.ent.cloudera.com:8080/1158
Tested-by: jenkins
Reviewed-by: Matteo Bertozzi <matteo.bertozzi@cloudera.com>