Add OVERRIDE and FINAL support to port.h
This adds defines for these extensions which are available in clang.
These keywords are C++11 keywords, but are supported by clang even if
c++11 mode isn't enabled, since they're back-compattible.

The ifdefs are borrowed from Chromium's compiler_specific.h.

This patch only adds usage of OVERRIDE in a couple of places to make sure
that it builds properly on all of our slaves.

In the future we can add some kind of clang plugin which verifies that the
OVERRIDE keyword is present on every instance of an overriding function.

Change-Id: I314b82431aae77c3a4492fc2ecf8ef288b8b68f6
Reviewed-on: http://gerrit.ent.cloudera.com:8080/1149
Tested-by: jenkins
Reviewed-by: Michael Percy <mpercy@cloudera.com>