Add a LatchCallback so that we can wait for async methods that receive callbacks
In some situations, e.g. tests, booting, we need to wait for certain methods to
complete which would normally run asynchronously. Such asynchronous methods usually
take callbacks. This adds a new callback that contains a latch and can be passed
to such methods and can then be Wait()ed upon.

Change-Id: Ic40ae45b8fda86cbb07f3aa0b86d4de1e1a310e3
Reviewed-on: http://gerrit.ent.cloudera.com:8080/1176
Tested-by: jenkins
Reviewed-by: Matteo Bertozzi <matteo.bertozzi@cloudera.com>