rpc: change DCHECK to CHECK that Reactor isn't shut down from a Reactor thread
It's relatively easily to accidentally shut down a Messenger from the reactor
thread - for example when a user's RPC callback ends up dropping a reference
to the Messenger. This isn't a hot path, so we should CHECK here rather than
risking accidentally getting stuck in a thread::join on the same thread.

A previous revision of this patch attempted to allow this behavior by deferring
the messenger shutdown to a separate thread. This solved the immediate problem,
but had the potential to introduce leak checker failures, since the deleter
thread might not have run by the time the program exited. So, it's safer to
just force the user to hang onto the Messenger, and trigger an assertion if
they don't do so.

Change-Id: Id5f6c3c82dbec51218a869c38d078982b12a0bca
Reviewed-on: http://gerrit.ent.cloudera.com:8080/1192
Tested-by: jenkins
Reviewed-by: Jean-Daniel Cryans <jdcryans@cloudera.com>