KUDU-29. Add a Clang-based tool to find cases where Status results are ignored
This uses the Clang ASTMatcher facility to find cases where we have
an expression of type Status directly inside a compound statement. In
other words, cases where we call a function which returns Status,
but don't do anything with it.

Instructions for compilation and usage are included in the README.
This is built against Clang 3.4 which isn't released quite yet, but
should be any day now. Once it is released and part of our toolchain,
we can enable this as a precommit check.

Change-Id: Id6847ad0a7d223ed9a2deac7dbed5a6912a3ff8c
Reviewed-on: http://gerrit.ent.cloudera.com:8080/1197
Tested-by: jenkins
Reviewed-by: David Alves <david.alves@cloudera.com>
Reviewed-by: Alex Feinberg <afeinberg@cloudera.com>