Fix Messenger to allow shutdown from within Reactor threads
After many different iterations of this patch, the final solution is
to change the teardown of Messenger such that the Reactor threads
actually retain shared ownership of the Messenger. This allows
the Reactor threads to shut down asynchronously with regard to
the user's Shutdown() call, since they know that the Messenger
object will not be destroyed.

To keep the behavior such that the user can shutdown the Messenger
by just dropping the shared_ptr reference, an second shared_ptr
is created with a custom deleter implementation that calls
Messenger::Shutdown().

A full explanation of the new ownership graph is provided in the
comments.

Change-Id: I4145058dcc504a787be2d3ee26679ce981306c71
Reviewed-on: http://gerrit.ent.cloudera.com:8080/1212
Tested-by: jenkins
Reviewed-by: Michael Percy <mpercy@cloudera.com>